[ci skip] Move splitter and data-writing functionality from MRT to CDS
